+++ crates/biome_js_analyze/src/lint/nursery.rs
@@ -15,8 +15,8 @@ pub mod no_namespace_import;
 pub mod no_nodejs_modules;
 pub mod no_re_export_all;
 pub mod no_restricted_imports;
-pub mod no_semicolon_in_jsx;
 pub mod no_skipped_tests;
+pub mod no_suspicious_semicolon_in_jsx;
 pub mod no_undeclared_dependencies;
 pub mod no_useless_ternary;
 pub mod use_import_restrictions;
@@ -41,8 +41,8 @@ declare_group! {
             self :: no_nodejs_modules :: NoNodejsModules ,
             self :: no_re_export_all :: NoReExportAll ,
             self :: no_restricted_imports :: NoRestrictedImports ,
-            self :: no_semicolon_in_jsx :: NoSemicolonInJsx ,
             self :: no_skipped_tests :: NoSkippedTests ,
+            self :: no_suspicious_semicolon_in_jsx :: NoSuspiciousSemicolonInJsx ,
             self :: no_undeclared_dependencies :: NoUndeclaredDependencies ,
             self :: no_useless_ternary :: NoUselessTernary ,
             self :: use_import_restrictions :: UseImportRestrictions ,