+++ crates/biome_js_analyze/src/lint/nursery/no_suspicious_semicolon_in_jsx.rs
@@ -42,14 +42,14 @@ declare_rule! {
     /// }
     /// ```
     ///
-    pub NoSemicolonInJsx {
+    pub NoSuspiciousSemicolonInJsx {
         version: "1.6.0",
-        name: "noSemicolonInJsx",
+        name: "noSuspiciousSemicolonInJsx",
         recommended: true,
     }
 }
 
-impl Rule for NoSemicolonInJsx {
+impl Rule for NoSuspiciousSemicolonInJsx {
     type Query = Ast<AnyJsxElement>;
     type State = TextRange;
     type Signals = Option<Self::State>;